% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{gen_flux_canny_pro1}
\alias{gen_flux_canny_pro1}
\title{Generate an image with FLUX.1 Canny pro using a control image}
\usage{
gen_flux_canny_pro1(
  control_image,
  prompt = NULL,
  steps = NULL,
  prompt_upsampling = NULL,
  seed = NULL,
  guidance = NULL,
  output_format = NULL,
  saftey_tolerance = NULL,
  output = c("result", "response", "request"),
  api_key = Sys.getenv("BFL_API_KEY"),
  open_with = "magick",
  download_path = NULL
)
}
\arguments{
\item{control_image}{Required. Base64 encoded image to use as control input.}

\item{prompt}{A character string representing the text prompt for image generation. This should
describe the desired image in detail.}

\item{steps}{Integer. Number of steps for the image generation process between 1 and 50.}

\item{prompt_upsampling}{A boolean indicating whether to perform upsampling on the prompt.
Default is FALSE.}

\item{seed}{An optional seed for reproducibility of the generated image. Default is NULL.}

\item{guidance}{Guidance strength for the image generation process. Number between 1 and 100.}

\item{output_format}{A character string specifying the output format of the generated image. Can
be 'jpeg' or 'png'. Default is NULL.}

\item{output}{Determines whether the function returns the request, response, or result. If}

\item{open_with}{Either 'none', 'browser', or 'magick'. Will look for \code{bfl_open_with} option
and default to 'magick' if not found.}

\item{download_path}{If \code{TRUE} and output = 'result', then passing a path to this argument will
download the file to that specified path.}
}
\value{
Either an \code{httr2} response or request object, or the JSON body of the finished result.
}
\description{
Submits an image generation task with FLUX.1 Canny pro.
}
\examples{
\dontrun{
 control_image <- base64enc::base64encode(system.file("/images/low-res-pink-hair-figurine.jpeg"))
 bfl::gen_flux_canny_pro1(
   control_image = control_image,
   prompt = paste0(
     "small humanoid creature with pink hair gently grasped by a human thumb and three fingers, ",
     "mars desert and sand dunes in the background, ultrarealistic, detailed"
   ),
   seed = 2,
   steps = 50,
   prompt_upsampling = TRUE,
   download_path = paste0("flux-canny-img_", Sys.time(), ".jpg")
 )
}
}
