% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{gen_flux_depth_pro1}
\alias{gen_flux_depth_pro1}
\title{Generate an image with FLUX.1 Depth pro using a control image.}
\usage{
gen_flux_depth_pro1(
  control_image,
  prompt = NULL,
  steps = NULL,
  prompt_upsampling = NULL,
  seed = NULL,
  guidance = NULL,
  output_format = NULL,
  saftey_tolerance = NULL,
  output = c("request", "response", "result"),
  api_key = Sys.getenv("BFL_API_KEY"),
  follow_url = TRUE,
  download_path = NULL
)
}
\arguments{
\item{control_image}{Required. Base64 encoded image to use as control input.}

\item{prompt}{A character string representing the text prompt for image generation. This should
describe the desired image in detail.}

\item{steps}{Integer. Number of steps for the image generation process between 1 and 50.}

\item{prompt_upsampling}{A boolean indicating whether to perform upsampling on the prompt.
Default is FALSE.}

\item{seed}{An optional seed for reproducibility of the generated image. Default is NULL.}

\item{guidance}{Guidance scale for image generation. High guidance scales improve prompt
adherence at the cost of reduced realism. Number between 1.5 and 5.}

\item{output_format}{A character string specifying the output format of the generated image. Can
be 'jpeg' or 'png'. Default is NULL.}

\item{output}{Determines whether the function returns the request, response, or result. If}

\item{follow_url}{Logical. If \code{TRUE} and output = 'result', then the generated image will in your
browser.}

\item{download_path}{If \code{TRUE} and output = 'result', then passing a path to this argument will
download the file to that specified path.}
}
\value{
Either an \code{httr2} response or request object, or the JSON body of the finished result.
}
\description{
Submits an image generation task with FLUX.1 Depth pro.
}
\examples{
TODO:
\dontrun{
gen_flux_depth_pro1(
  control_image =
  prompt = "A serene landscape with mountains in the background and a calm lake in the foreground.",
  seed = 1
 )
}
}
