% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{gen_flux_fill_pro1}
\alias{gen_flux_fill_pro1}
\title{Generate an Image with FLUX 1.0 Fill Pro}
\usage{
gen_flux_fill_pro1(
  image,
  mask = NULL,
  prompt = NULL,
  steps = NULL,
  prompt_upsampling = NULL,
  seed = NULL,
  guidance = NULL,
  output_format = NULL,
  saftey_tolerance = NULL,
  output = c("result", "response", "request"),
  api_key = Sys.getenv("BFL_API_KEY"),
  follow_url = TRUE,
  download_path = NULL
)
}
\arguments{
\item{image}{Required. A Base64-encoded string representing the image you wish to modify. Can
contain alpha mask if desired.}

\item{mask}{A Base64-encoded string representing a mask for the areas you want to modify in the
image. The mask should be the same dimensions as the image and in black and white. Black areas
(0\%) indicate no modification, while white areas (100\%) specify areas for inpainting. Optional
if you provide an alpha mask in the original image. Validation: The endpoint verifies that the
dimensions of the mask match the original image.}

\item{prompt}{The description of the changes you want to make. This text guides the inpainting process,
allowing you to specify features, styles, or modifications for the masked area.}

\item{steps}{Integer. Number of steps for the image generation process between 1 and 50.}

\item{prompt_upsampling}{A boolean indicating whether to perform upsampling on the prompt.
Default is FALSE.}

\item{seed}{An optional seed for reproducibility of the generated image. Default is NULL.}

\item{guidance}{Guidance scale for image generation. High guidance scales improve prompt
adherence at the cost of reduced realism. Number between 1.5 and 5.}

\item{output_format}{A character string specifying the output format of the generated image. Can
be 'jpeg' or 'png'. Default is NULL.}

\item{output}{Determines whether the function returns the request, response, or result. If}

\item{follow_url}{Logical. If \code{TRUE} and output = 'result', then the generated image will in your
browser.}

\item{download_path}{If \code{TRUE} and output = 'result', then passing a path to this argument will
download the file to that specified path.}
}
\value{
Either an \code{httr2} response or request object, or the JSON body of the finished result.
}
\description{
Submits an image generation task with the FLUX.1 Fill pro model using an input image and mask.
Mask can be applied to alpha channel or submitted as a separate image.
}
\examples{
\dontrun{
mask <- base64enc::base64encode(system.file("/images/IMG_3130_mask.jpeg", package = "bfl"))
image <- base64enc::base64encode(system.file("/images/IMG_3130.jpeg", package = "bfl"))
gen_flux_fill_pro1(
image = image,
  mask = mask,
  prompt = "alien spaceship, ultrarealistic, detailed",
  seed = 200,
  steps = 50,
  prompt_upsampling = TRUE,
  download_path = paste0(Sys.time(), ".jpg")
)
}
}
