% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{gen_flux_pro}
\alias{gen_flux_pro}
\title{Generate an Image with FLUX Pro}
\usage{
gen_flux_pro(
  prompt,
  image_prompt = NULL,
  width = NULL,
  height = NULL,
  steps = NULL,
  prompt_upsampling = NULL,
  seed = NULL,
  guidance = NULL,
  saftey_tolerance = NULL,
  interval = NULL,
  output_format = NULL,
  output = c("request", "response", "result"),
  api_key = Sys.getenv("BFL_API_KEY"),
  follow_url = TRUE,
  download_path = NULL
)
}
\arguments{
\item{prompt}{A character string representing the text prompt for image generation. This should
describe the desired image in detail.}

\item{image_prompt}{An optional base64 encoded image that can be used with Flux Redux. Default is
NULL.}

\item{width}{An integer specifying the width of the generated image in pixels. Must be between
256 and 1440, and a multiple of 32. Default is 1024.}

\item{height}{An integer specifying the height of the generated image in pixels. Must be between
256 and 1440, and a multiple of 32. Default is 768.}

\item{steps}{Integer. Number of steps for the image generation process between 1 and 50.}

\item{prompt_upsampling}{A boolean indicating whether to perform upsampling on the prompt.
Default is FALSE.}

\item{seed}{An optional seed for reproducibility of the generated image. Default is NULL.}

\item{guidance}{Guidance scale for image generation. High guidance scales improve prompt
adherence at the cost of reduced realism. Number between 1.5 and 5.}

\item{interval}{Interval parameter for guidance control. Number between 1 and 4.}

\item{output_format}{A character string specifying the output format of the generated image. Can
be 'jpeg' or 'png'. Default is NULL.}

\item{output}{Determines whether the function returns the request, response, or result. If}

\item{follow_url}{Logical. If \code{TRUE} and output = 'result', then the generated image will in your
browser.}

\item{download_path}{If \code{TRUE} and output = 'result', then passing a path to this argument will
download the file to that specified path.}

\item{safety_tolerance}{An integer between 0 and 6 specifying the tolerance level for input and
output moderation. 0 is most strict, and 6 is least strict. Default is NULL.}
}
\value{
Either an \code{httr2} response or request object, or the JSON body of the finished result.
}
\description{
Submits a request to generate an image using the FLUX 1.1 \link{Pro} API endpoint.
The function allows customization of the image generation through several parameters such as prompt,
image dimensions, and other options like upsampling, seed, and output format.
}
\examples{
\dontrun{
gen_flux_pro(
  prompt = "A serene landscape with mountains in the background and a calm lake in the foreground.",
  seed = 1
 )
}
}
