% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{gen_flux_pro1.1_ultra}
\alias{gen_flux_pro1.1_ultra}
\title{Generate an Image with FLUX 1.1 Pro Ultra}
\usage{
gen_flux_pro1.1_ultra(
  prompt,
  seed = NULL,
  aspect_ratio = NULL,
  saftey_tolerance = NULL,
  output_format = NULL,
  raw = NULL,
  image_prompt = NULL,
  image_prompt_strength = NULL,
  output = c("result", "response", "request"),
  api_key = Sys.getenv("BFL_API_KEY"),
  follow_url = TRUE,
  download_path = NULL
)
}
\arguments{
\item{prompt}{A character string representing the text prompt for image generation. This should
describe the desired image in detail.}

\item{seed}{An optional seed for reproducibility of the generated image. Default is NULL.}

\item{aspect_ratio}{Aspect ratio of the image between 21:9 and 9:21. Default is 16:9.}

\item{output_format}{A character string specifying the output format of the generated image. Can
be 'jpeg' or 'png'. Default is NULL.}

\item{raw}{Generate less processed, more natural-looking images. Default is \code{FALSE}.}

\item{image_prompt}{An optional base64 encoded image that can be used with Flux Redux. Default is
NULL.}

\item{image_prompt_strength}{Blend between the prompt and the image prompt. Number between 0 and 1.
Default is 0.1.}

\item{output}{Determines whether the function returns the request, response, or result. If}

\item{follow_url}{Logical. If \code{TRUE} and output = 'result', then the generated image will in your
browser.}

\item{download_path}{If \code{TRUE} and output = 'result', then passing a path to this argument will
download the file to that specified path.}
}
\value{
Either an \code{httr2} response or request object, or the JSON body of the finished result.
}
\description{
Submits an image generation task with FLUX 1.1 pro with ultra mode and optional raw mode.
}
\examples{
\dontrun{
  gen_flux_pro1.1_ultra(
    prompt = "A serene landscape with mountains in the background and a calm lake in the foreground.",
    seed = 1,
    raw = TRUE
   )
}
}
