% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-result.R
\name{get_result}
\alias{get_result}
\title{Get Result from BFL API}
\usage{
get_result(
  id,
  x_key = Sys.getenv("BFL_API_KEY"),
  output = c("json", "response", "request")
)
}
\arguments{
\item{id}{A character string representing the unique identifier for the result to be retrieved.}

\item{x_key}{API key. By default, this is retrieved from the environment variable \code{BFL_API_KEY}.}
}
\value{
The response object from the API request.
}
\description{
This function sends a GET request to the BFL API to retrieve the result associated with a specific ID.
}
\examples{
\dontrun{
# Example usage of the function
result <- get_result(id = "12345")
}

}
