% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-result.R
\name{get_result_status}
\alias{get_result_status}
\title{Get Result Status}
\usage{
get_result_status(id, x_key = Sys.getenv("BFL_API_KEY"))
}
\arguments{
\item{id}{A character string representing the unique identifier for the result to be retrieved.}

\item{x_key}{API key. By default, this is retrieved from the environment variable \code{BFL_API_KEY}.}
}
\description{
Get Result Status
}
