% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_rows.R
\name{hash_rows}
\alias{hash_rows}
\title{Hash rows of a matrix or data.frame}
\usage{
hash_rows(X, algo = "xxhash64")
}
\arguments{
\item{X}{Matrix or data.frame}

\item{algo}{Hash algorithm}
}
\description{
Internal helper for row-wise alignment of target datasets.
}
\keyword{internal}
