% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVA_helpers.R
\name{LCVA_interface_layer}
\alias{LCVA_interface_layer}
\title{LCVA Interface Layer}
\description{
Clean interface around \pkg{LCVA} for fitting and predicting
on single or multiple datasets.
}
\keyword{internal}
