% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_BFL_deterministic.R
\name{predict_BFL_deterministic}
\alias{predict_BFL_deterministic}
\title{Deterministic BFL prediction (posterior-mean probabilities)}
\usage{
predict_BFL_deterministic(global_fit, return_probs = TRUE)
}
\description{
Deterministic BFL prediction (posterior-mean probabilities)
}
\keyword{internal}
