% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-result.R
\name{repeat_until}
\alias{repeat_until}
\title{Repeat a function until it returns expected value}
\usage{
repeat_until(
  func,
  expected_result,
  failed_result = NULL,
  sleep_time = 1,
  timeout = 60,
  ...
)
}
\arguments{
\item{func}{Function to run.}

\item{expected_result}{Expected result of \code{func} that ends loop of repeated evaluation.}

\item{sleep_time}{Optional amount of seconds to sleep between retries.}

\item{...}{Other arguments passed to func.}
}
\description{
Repeat a function until it returns expected value
}
