% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_BFL.R
\name{summary_BFL}
\alias{summary_BFL}
\title{Scientist-facing summary for BFL predictions}
\usage{
summary_BFL(pred_out, y_true, csmf_fun = csmf_accuracy, cause_level = FALSE)
}
\arguments{
\item{pred_out}{Output of predict_BFL(return_probs = TRUE)}

\item{y_true}{Character vector of true causes (length N)}

\item{csmf_fun}{Function(pi_hat, pi_true) -> numeric}

\item{cause_level}{Logical; if TRUE, compute per-cause table (slow)}
}
\description{
Scientist-facing summary for BFL predictions
}
