% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_causes.R
\name{align_local_summaries}
\alias{align_local_summaries}
\title{Align local summaries to a global cause set (internal)}
\usage{
align_local_summaries(local_summaries, ref_row_hash = NULL)
}
\arguments{
\item{local_summaries}{list of local summary objects}

\item{ref_row_hash}{optional reference hash for row alignment}
}
\value{
list(global_causes, aligned_phi, ref_row_hash)
}
\description{
Align local summaries to a global cause set (internal)
}
\keyword{internal}
