% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_BFL.R
\name{predict_BFL}
\alias{predict_BFL}
\title{Predict causes using a global BFL fit}
\usage{
predict_BFL(
  global_fit,
  sampling = FALSE,
  ndraws = NULL,
  seed = NULL,
  return_probs = TRUE
)
}
\arguments{
\item{global_fit}{Output of run_BFL()}

\item{sampling}{Logical; if TRUE, use posterior predictive sampling.}

\item{ndraws}{Number of posterior draws to use when sampling (NULL = all).}

\item{seed}{Optional seed for reproducibility (only used when sampling = TRUE or ndraws not NULL).}

\item{return_probs}{Logical; return posterior-mean probabilities.}
}
\description{
Predict causes using a global BFL fit
}
